Clean up withAuthHeader
